+++ src/common/components/bitcoin-links/block-hash.tsx
@@ -1,6 +1,6 @@
 import React from 'react';
 import * as constants from '../../../constants';
-import InterlayLink from 'components/InterlayLink';
+import InterlayLink from 'components/UI/InterlayLink';
 
 export default class BitcoinBlockHash extends React.Component<{
   blockHash: string;
