+++ src/pages/Challenges/index.tsx
@@ -9,17 +9,17 @@ import PageTitle from 'parts/PageTitle';
 import TimerIncrement from 'parts/TimerIncrement';
 import VaultScoresTable from 'containers/VaultScoresTable';
 import StakedRelayerScoresTable from 'containers/StakedRelayerScoresTable';
-import InterlayLink from 'components/InterlayLink';
-import InterlayTabs, { InterlayTab } from 'components/InterlayTabs';
-import InterlayToggleButtonGroup, {
-  InterlayToggleButton,
-  InterlayToggleButtonGroupProps
-} from 'components/InterlayToggleButtonGroup';
 import CardList, {
   Card,
   CardHeader,
   CardContent
 } from 'components/CardList';
+import InterlayLink from 'components/UI/InterlayLink';
+import InterlayTabs, { InterlayTab } from 'components/UI/InterlayTabs';
+import InterlayToggleButtonGroup, {
+  InterlayToggleButton,
+  InterlayToggleButtonGroupProps
+} from 'components/UI/InterlayToggleButtonGroup';
 import {
   POLKA_BTC_DOC_TREASURE_HUNT,
   POLKA_BTC_DOC_TREASURE_HUNT_VAULT,
